[FE] move: BaseModal common 하위로 이동 #115
