+++ frontend/src/components/Modals/common/BaseModal.vue
